requirements.txt
Django
django-crispy-forms
crispy-bootstrap4
crispy-bootstrap5
django-bootstrap4
django-bootstrap-v5
openpyxl
django-import-export
Pillow
python-barcode
